Split make_sub_folders and formatting tests into two modules.
